Move eth2 code inside the eth2 module
Fix #2719
